+++ fragments/labels/acroniscyberprotectconnectagent.sh
@@ -7,4 +7,5 @@ remotixagent)
     versionKey="CFBundleVersion"
     appNewVersion=$(curl -fsIL "${downloadURL}" | grep -i "^location" | sed -E 's/.*\/[a-zA-Z]*-[0-9.]*-([0-9.]*)\.pkg/\1/g')
     expectedTeamID="H629V387SY"
+    blockingProcesses=( NONE )
     ;;